feat(kanban): unify logic in useJobBoard, extract UI, fix tests
